+++ R/model_fertility.R
@@ -39,8 +39,8 @@
 #'   * von Bertalanffy: \eqn{f(x) = A  (1 - exp(-k  (x - x_0)))}
 #'   * Normal: \eqn{f(x) = A \times \exp\left(
 #'   -\frac{1}{2}\left(\frac{x-\mu}{\sigma}\right)^{\!2}\,\right)}
-#'   * Hadwiger: \eqn{f(x) = \frac{ab}{c} \left (\frac{c}{x}  \right )
-#'    ^\frac{3}{2} \exp \left \{ -b^2  \left ( \frac{c}{x}+\frac{x}{c}-2
+#'   * Hadwiger: \eqn{f(x) = \frac{ab}{C} \left (\frac{C}{x}  \right )
+#'    ^\frac{3}{2} \exp \left \{ -b^2  \left ( \frac{C}{x}+\frac{x}{C}-2
 #'    \right ) \right \}}
 #'
 #'
@@ -75,13 +75,13 @@
 #'
 #' # Compute fertility using the Hadwiger model
 #' model_fertility(
-#'   age = 0:50, params = c(a = 0.91, b = 3.85, c = 29.78),
+#'   age = 0:50, params = c(a = 0.91, b = 3.85, C = 29.78),
 #'   maturity = 0, model = "hadwiger"
 #' )
 #'
 #' @family trajectories
 #' @author Owen Jones <jones@biology.sdu.dk>
-#' @seealso [model_survival()] to model age-specific survival using mortality
+#' @seealso [model_mortality()] to model age-specific survival using mortality
 #'   models.
 #' @export